requirements.txt
requests
colorlog
websockets
lxml
aiohttp
pyyaml
